requirements.txt
biopython==1.79
pandas~=1.4.2
joblib~=1.1.0